Acikgoz/requirements.txt
torch~=1.13.1
pillow~=8.0.1
torchvision~=0.14.1
yaml~=0.2.5
pyyaml~=5.3.1